has_permissions_on_user should ignore subject's inactive assignments This could be simplified by adding an association to the User model such as:
```ruby
has_many :active_units, -> { Assignment.active }, through: assignments, class: 'User', source: :unit
```

Then we could go back to a simple `subject.units.flat_map(&:path_ids).uniq`. But what to name that association?
Calling it 'active_units' implies it's filtering on `units.active`, but it would actually be filtering on the assignments
through which it reaches the units. So I opted to simply make the logic explicit in this permissions query until we
need to re-use it elsewhere, at which point it probably makes sense to add an association like this.